Bug Fixes in models and db-schemas

+++ database/migrations/2017_01_28_073511_create_camps_table.php
@@ -17,8 +17,8 @@ class CreateCampsTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('address');
-            $table->decimal('longitude')->nullable();
-            $table->decimal('latitude')->nullable();
+            $table->float('longitude')->nullable();
+            $table->float('latitude')->nullable();
             $table->string('code', 10);
             $table->unsignedInteger('lga_id');
             $table->nullableTimestamps();
